requirements-agent.txt
# Agent dependencies (on top of requirements.txt base)
fastapi>=0.135.0
uvicorn[standard]>=0.43.0
websockets>=14.0
litellm>=1.82.3
langchain>=1.2.13
langchain-core>=1.2.22
langchain-community>=0.4.1
langchain-litellm>=0.6.2
langgraph>=1.1.3
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1.0
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
pyyaml>=6.0
# observability
langfuse>=4.0.1
# recall.ai integration (HTTP client)
httpx>=0.27.0
# local TTS backend
kokoro-onnx>=0.5.0
# voice activity detection for STT pipeline
webrtcvad>=2.0.10
